+++ src/RcppExports.cpp
@@ -86,25 +86,27 @@ BEGIN_RCPP
 END_RCPP
 }
 // split_frequencies
-List split_frequencies(const List trees);
-RcppExport SEXP _TreeTools_split_frequencies(SEXP treesSEXP) {
+List split_frequencies(const List trees, const bool exact);
+RcppExport SEXP _TreeTools_split_frequencies(SEXP treesSEXP, SEXP exactSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const List >::type trees(treesSEXP);
-    rcpp_result_gen = Rcpp::wrap(split_frequencies(trees));
+    Rcpp::traits::input_parameter< const bool >::type exact(exactSEXP);
+    rcpp_result_gen = Rcpp::wrap(split_frequencies(trees, exact));
     return rcpp_result_gen;
 END_RCPP
 }
 // consensus_tree
-RawMatrix consensus_tree(const List trees, const NumericVector p);
-RcppExport SEXP _TreeTools_consensus_tree(SEXP treesSEXP, SEXP pSEXP) {
+RawMatrix consensus_tree(const List trees, const NumericVector p, const bool exact);
+RcppExport SEXP _TreeTools_consensus_tree(SEXP treesSEXP, SEXP pSEXP, SEXP exactSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const List >::type trees(treesSEXP);
     Rcpp::traits::input_parameter< const NumericVector >::type p(pSEXP);
-    rcpp_result_gen = Rcpp::wrap(consensus_tree(trees, p));
+    Rcpp::traits::input_parameter< const bool >::type exact(exactSEXP);
+    rcpp_result_gen = Rcpp::wrap(consensus_tree(trees, p, exact));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -541,8 +543,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_TreeTools_ape_neworder_phylo", (DL_FUNC) &_TreeTools_ape_neworder_phylo, 5},
     {"_TreeTools_ape_neworder_pruningwise", (DL_FUNC) &_TreeTools_ape_neworder_pruningwise, 5},
     {"_TreeTools_as_newick", (DL_FUNC) &_TreeTools_as_newick, 1},
-    {"_TreeTools_split_frequencies", (DL_FUNC) &_TreeTools_split_frequencies, 1},
-    {"_TreeTools_consensus_tree", (DL_FUNC) &_TreeTools_consensus_tree, 2},
+    {"_TreeTools_split_frequencies", (DL_FUNC) &_TreeTools_split_frequencies, 2},
+    {"_TreeTools_consensus_tree", (DL_FUNC) &_TreeTools_consensus_tree, 3},
     {"_TreeTools_descendant_edges", (DL_FUNC) &_TreeTools_descendant_edges, 3},
     {"_TreeTools_descendant_edges_single", (DL_FUNC) &_TreeTools_descendant_edges_single, 5},
     {"_TreeTools_descendant_tips", (DL_FUNC) &_TreeTools_descendant_tips, 3},